restructure UAM code into separate files
uams.c was getting unweildy with all UAM code in one place;
this splits it up by functional UAM, except for Guest auth
which is trivial for the client
